Move login button from layout to page
